Fix single-order logic
The previous solution relied on cancelling orders from the frontend whenever a new order is entered. This created a bug where orders could be entered before a previous order's confirmation was recieved by the frontend. This means the order would be entered without the previous order being canceled.

To solve this, I just copied my strategy from visual markets. The single-order logic is now implemented on the backend, by canceling orders when their confirmations return from the exchange. This is much more robust and should avoid multiple-order-entry bugs such as this one.